requirements.txt
streamlit==1.41.1
langchain-text-splitters==0.3.5
langchain-community==0.3.15
langgraph==0.2.67
PyPDF2==3.0.1
langchain-openai==0.3.2
langchain-community==0.3.15
pypdf==5.1.0
python-dotenv==1.0.1
langchain-groq==0.2.3
langchain-google-genai==2.0.9
